+++ src/components/shared/images/upload/ImageUploadEntries.tsx
@@ -1,25 +1,10 @@
 import React from 'react'
 import useImageUpload from '../../../../hooks/use-image-upload'
 import { IconButton, ImageList, ImageListItem, ImageListItemBar, Tooltip } from '@mui/material'
-import { Delete, LocationOn, LocationOff } from '@mui/icons-material'
+import { Delete, LocationOff, LocationOn } from '@mui/icons-material'
 import Typography from '@mui/material/Typography'
 import getCloudStorageUrlForIdentifier from '../../../../utils/storage-helper'
-
-type GeoReferenceIndicatorProps = {
-  isGeoReferenced: boolean
-}
-
-const GeoReferenceIndicator = ({ isGeoReferenced }:GeoReferenceIndicatorProps) => {
-  const title = isGeoReferenced ? 'Coordinates found' : 'No coordinates found'
-  return (
-    <Tooltip title={title}>
-      {isGeoReferenced
-        ? <LocationOn sx={{ color: 'rgba(255, 255, 255, 0.54)' }} />
-        : <LocationOff sx={{ color: 'rgba(255, 255, 255, 0.54)' }} />
-      }
-    </Tooltip>
-  )
-}
+import GeoReferenceIndicator from './GeoReferenceIndicator'
 
 const ImageUploadEntries = () => {
   const { files, remove } = useImageUpload()
@@ -34,8 +19,8 @@ const ImageUploadEntries = () => {
         Currently uploaded images:
       </Typography>
       <Typography variant="caption" component="div">
-        Geo-referenced images will be placed on the map directly. Successful coordinate extraction is marked
-        with  <LocationOn fontSize={'inherit'}/>.
+        Geo-referenced images will be placed on the map directly. Images with successfully extracted coordinates are
+        marked with <LocationOn fontSize={'inherit'}/>, otherwise with <LocationOff fontSize={'inherit'}/>.
       </Typography>
       <ImageList cols={6} gap={8}>
         {files.map((item, index) => (
@@ -46,16 +31,16 @@ const ImageUploadEntries = () => {
               loading="lazy"
             />
             <ImageListItemBar
-              subtitle={item.location ? <GeoReferenceIndicator isGeoReferenced /> : <GeoReferenceIndicator isGeoReferenced={false}/>}
+              subtitle={<GeoReferenceIndicator isGeoReferenced={!!item.location}/>}
               actionIcon={
-              <Tooltip title={'Delete image'}>
-                <IconButton
-                  sx={{ color: 'rgba(255, 255, 255, 0.54)' }}
-                  onClick={() => remove(item.id)}
-                >
-                  <Delete/>
-                </IconButton>
-              </Tooltip>
+                <Tooltip title={'Delete image'}>
+                  <IconButton
+                    sx={{ color: 'rgba(255, 255, 255, 0.54)' }}
+                    onClick={() => remove(item.id)}
+                  >
+                    <Delete/>
+                  </IconButton>
+                </Tooltip>
               }
             />
           </ImageListItem>
